requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
supabase==2.28.0
redis==7.3.0
celery==5.6.2
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
aiofiles==25.1.0
httpx==0.28.1
elasticsearch==8.11.0
pika==1.3.2
boto3==1.42.65
python-dotenv==1.2.2
email-validator==2.3.0
requests==2.32.5
resend==2.7.0